Moved get_constance_field/set_constance_value to forms.py
