+++ constance/utils.py
@@ -3,8 +3,9 @@ from __future__ import unicode_literals
 
 from importlib import import_module
 
-from constance import config, settings
-from constance.forms import ConstanceForm
+from constance import LazyConfig, settings
+
+config = LazyConfig()
 
 
 def import_module_attr(path):
@@ -24,22 +25,3 @@ def get_constance_values():
                    **dict(config._backend.mget(settings.CONFIG.keys())))
 
     return initial
-
-
-def get_constance_field(name):
-    form = ConstanceForm(initial=get_constance_values())
-
-    return form.fields[name]
-
-
-def set_constance_value(raw_name, raw_value):
-    """
-    Parses and sets a Constance value
-    :param raw_name:
-    :param raw_value:
-    :return:
-    """
-    field = get_constance_field(raw_name)
-
-    value = field.clean(field.to_python(raw_value))
-    setattr(config, raw_name, value)
